try to fix issues with seconds changes on a vt240
Just resurrected an old DEC VT240 (the color model so it's really a
VT241) and it seems to be mostly working.  Mostly.  I've been seeing a
little choppiness on the clock updates.  Sometimes a second will skip
or a second count will display for 1.5 or 2 seconds or something like
that.

My thinking here is that gettimeofday() and time() might be out of
sync.  And it's probably a good idea to pick just one instead of using
the former during the delay loop and the latter to get the timestamp
to display anyway.

A previous attempt to use halfdelay(1) didn't fix this.  The premise
behind that was that a loop repeating calls to gettimeofday() and
wgetch() might be a little wasteful and may have been causing the
issue.  The halfdelay(1) call would simply add a slight delay in each
iteration of that loop.

There's a mydelay fn left over that for some reason I'm not calling
anymore.